+++ platform-shop/src/main/java/com/platform/dao/GoodsDao.java
@@ -10,5 +10,5 @@ import com.platform.entity.GoodsEntity;
  * @date 2017-08-21 21:19:49
  */
 public interface GoodsDao extends BaseDao<GoodsEntity> {
-    Integer queryMaxId();
+    Long queryMaxId();
 }
